+++ Assets/Images/GameAssets/vfx/Prefabs/Kunaitrail.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 418c7a7a2a1aa46f09c11da09410ad0d
+guid: 5d42a3f28c53442eba3ab534e0007a57
 PrefabImporter:
   externalObjects: {}
   userData:
